requirements_test.txt
pytest-homeassistant-custom-component==0.12.19

# Component dependencies
backoff~=2.0
aiolimiter~=1.0